Handle rebooting differently (again)
If --yes-to-all is provided then we will exit with code 2 signalling
that a reboot should be performed as soon as possible. This allows
surrounding automation to do more stuff before the reboot.

If --reboot is provided then we will do the reboot immediately ourselves.

Otherwise we will ask the user to do it for us.

Signed-off-by: David Scott <dave.scott@citrix.com>